Use fake time in counter tests
Unit tests that depend on real time are prone to spurious failures, as
the success of a check often winds up depending on the current load of
the host running the tests. This seems to affect TestCounterRate
frequently.

This "fixes" the problem for counters, which use a cached timestamp that
is asynchronously updated. Now, for tests, asynchronous updates stop and
time is advanced at will. It's a messy solution, but it fixes the unit
test without any significant changes to the main package.
